refactor(dashboard): extract hooks and components from SessionDetailPage
Split the 1077-line god component into focused modules:
- types.ts: TabId, ScreenshotState, COMMON_SLASH_COMMANDS
- useAuditData.ts: audit trail + PR entries fetching
- useMessageInput.ts: message sending with history navigation
- useScreenshot.ts: screenshot capture state management
- TabBar.tsx: tab navigation with animated indicator
- MessageFooter.tsx: desktop + mobile message input area

SessionDetailPage.tsx reduced to ~566 lines (composition + action handlers + tab panels).
Behavior identical — all 1096 tests pass, tsc --noEmit clean.
